filter is better but better cast to the right type

+++ schema/create-function-boiler_summary.sql
@@ -19,14 +19,14 @@ BEGIN
             event_group,
             system_status
         ) SELECT
-            SUM(cycles.btu) FILTER (WHERE cycles.system_status = 'boiler'),
-            SUM(cycles.gallons) FILTER (WHERE cycles.system_status = 'boiler'),
-            SUM(cycles.kwh),
-            COUNT(cycles.*) FILTER (WHERE cycles.system_status = 'boiler'),
-            SUM(cycles.runtime) FILTER (WHERE cycles.system_status = 'boiler'),
-            AVG(cycles.runtime) FILTER (WHERE cycles.system_status = 'boiler'),
-            MIN(cycles.runtime) FILTER (WHERE cycles.system_status = 'boiler'),
-            MAX(cycles.runtime) FILTER (WHERE cycles.system_status = 'boiler')
+            (SUM(cycles.btu) FILTER (WHERE cycles.system_status = 'boiler'))::NUMERIC,
+            (SUM(cycles.gallons) FILTER (WHERE cycles.system_status = 'boiler'))::NUMERIC,
+            (SUM(cycles.kwh))::NUMERIC,
+            (COUNT(cycles.*) FILTER (WHERE cycles.system_status = 'boiler'))::INTEGER,
+            (SUM(cycles.runtime) FILTER (WHERE cycles.system_status = 'boiler'))::NUMERIC,
+            (AVG(cycles.runtime) FILTER (WHERE cycles.system_status = 'boiler'))::NUMERIC,
+            (MIN(cycles.runtime) FILTER (WHERE cycles.system_status = 'boiler'))::NUMERIC,
+            (MAX(cycles.runtime) FILTER (WHERE cycles.system_status = 'boiler'))::NUMERIC
         FROM
             cycles;
 END;
